refactor: Split blockstore and datastore into their own modules.
