+++ src/repo/datastore/mod.rs
@@ -1,0 +1,3 @@
+pub mod sled;
+pub mod memory;
+pub mod flatfs;
